pyproject.toml
[tool.poetry]
name = "fastapi-keycloak"
version = "0.0.1"
description = ""
authors = ["Abdullah Alqahtani <abdullah@lameh.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
python-keycloak = "^4.4.0"
httpx = "^0.27.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
python-multipart = "^0.0.20"
uvicorn = "^0.34.0"
httptools = "^0.6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"